CA-281402: Update string fields to char ptrs
From the mailing list discussion:

> When the memory is owned by C code -- for example, when a C function
  passes a struct to OCaml with fields already initialized -- then using
  the string view works well: reading a string field creates an OCaml
  copy, which persists as long as needed.

> However, when things are the other way round -- i.e. when you're
  creating or initializing the struct in OCaml before passing it to C --
  then the string view isn't a good choice, because there's no way to
  manage the lifetime of the C copy of the OCaml string that's written
  to the struct field.

> A reasonable way to do this would be to allocate stable memory for the
  string fields before the C call, and ensure that the memory stays
  around until the call is complete.  Stable memory could mean any of: a
  bigarray, a Ctypes.CArray.t value, memory returned by malloc, memory
  returned by Ctypes.allocate, etc.  In most of these cases ensuring the
  memory stays around is a matter of holding onto the handle (bigarray
  value, Ctypes ptr, etc.) until after the call.

This PR is to remove the string view and use the char ptrs instead.
The tricky parts are extracting the string as we don't know the lenght,
and making them live long enough (but I think that was already ok).

Initially I did the following
```
(* We can't define it as
   CArray.(string_from_ptr (start a) (length a))
   as we do not know the length of the string a priori.

   Will fail with `Invalid_argument "index_out_of_bounds" if
   we don't reach the null terminator by [maxlen] chars.
*)
let char_array_as_string ?(maxlen=1024) a =
  (* here we overflow but we should never read after the string if the
     memory is not corrupted *)
  let a': char carray = CArray.from_ptr a maxlen in
  let b = Buffer.create 128 in
  let rec scan idx =
    let c = CArray.get a' idx in
    if c = '\x00' then
      Buffer.contents b
    else begin
      Buffer.add_char b c;
      scan (idx+1)
    end
  in
  scan 0
```

But I think that the above is morally equivalent to the current

```
let ptr_char_to_string p =
  let s = coerce (ptr char) string p in
  let b = Buffer.create 128 in
  Buffer.add_string b s;
  Buffer.contents b
```

although more explicit. In both cases we risk to do an overflow and
fail. It could be that the original is better as we fail more explicitly
while here we are somewhat hoping for the best. Either way, we are
strongly relying on the C strings to be null terminated and the memory
to be sane.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>