Fix lifetimes of the allocated memory
If we're allocating memory for use by c functions, we must make sure
that _all_ of the OCaml objects that are referencing the memory
live long enough. Particularly important, and what is fixed by this commit,
is that if you've got a structure with fields that are pointing at
allocated memory, it's not sufficient to keep the OCaml object that
references the structure alive - you _must_ keep all of the other
objects referncing the pointed-at memory in scope too - the GC isn't
omniscient!

The approach in this commit is to keep all the allocated objects together
in a tuple whose type is made opaque, and only provide accessors for
the values necessary (e.g. in this case, we need the address of the
structure).

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>